requirements.txt
azure-functions==1.18.0
numpy==1.26.3
pandas==2.2.0
psycopg2-binary==2.9.9
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
tzdata==2023.4

